Fix clang compilation
- Add macro for printing size_t independent of OS, processor bits (32/64)
- Deal with comparison of different types
- Add pose traits to their corresponding header files

+++ libs/graphslam/include/mrpt/graphslam/GSO/CLevMarqGSO.h
@@ -373,7 +373,7 @@ class CLevMarqGSO:
 		bool m_just_fully_optimized_graph;
 
 		/**\brief Minimum number of nodes before we try optimizing the graph */
-		int m_min_nodes_for_optimization;
+		size_t m_min_nodes_for_optimization;
 };
 
 } } } // end of namespaces
